moved status codes to enum
