+++ src/controllers/rate.controller.ts
@@ -1,14 +1,15 @@
-import BaseRateClient from '../services/rate/client';
-import CachedRateClient from '../services/rate/client.cache';
-import RateLogger from '../services/rate/logger';
+import BaseRateClient from '../services/rate/rate.client';
+import CachedRateClient from '../services/rate/rate.cache';
+import RateLogger from '../services/rate/rate.logger';
 import { Request, Response } from 'express';
+import { HttpCode } from '../http-responses/http-code.enum';
 
 export default async (req: Request, res: Response) => {
     try {
         const rateClient = new RateLogger(new CachedRateClient(new BaseRateClient()));
-        res.status(200).json(await rateClient.getRate());
+        res.status(HttpCode.OK).json(await rateClient.getRate());
     } catch (error: any) {
-        res.status(400).json({
+        res.status(HttpCode.BAD_REQUEST).json({
             message: error.message,
         });
     }
