+++ src/http-responses/http-code.enum.ts
@@ -1,0 +1,7 @@
+export enum HttpCode {
+    OK = 200,
+    BAD_REQUEST = 400,
+    NOT_FOUND = 404,
+    CONFLICT = 409,
+    INTERNAL_SERVER_ERROR = 500,
+}
